+++ src/python/dxpy/scripts/dx.py
@@ -100,8 +100,6 @@ if '_ARGCOMPLETE' not in os.environ:
         if old_term_setting:
             os.environ['TERM'] = old_term_setting
 
-        if readline.__doc__ and 'libedit' in readline.__doc__:
-            print('Warning: incompatible readline module detected (libedit), tab completion disabled', file=sys.stderr)
     except ImportError:
         if os.name != 'nt':
             print('Warning: readline module is not available, tab completion disabled', file=sys.stderr)
@@ -2752,8 +2750,28 @@ def render_timestamp(epochSeconds):
 
 def list_database_files(args):
     try:
+        # check if database was given as an object hash id
+        if is_hashid(args.database):
+            desc = dxpy.api.database_describe(args.database)
+            entity_result = {"id": desc["id"], "describe": desc}
+        else:
+        # otherwise it was provided as a path, so try and resolve
+            project, _folderpath, entity_result = try_call(resolve_existing_path,
+                                                           args.database,
+                                                           expected='entity')
+
+        # if we couldn't resolved the entity, fail
+        if entity_result is None:
+            err_exit('Could not resolve ' + args.database + ' to a data object', 3)
+        else:
+        # else check and verify that the found entity is a database object
+            entity_result_class = entity_result['describe']['class']
+            if entity_result_class != 'database':
+                err_exit('Error: The given object is of class ' + entity_result_class +
+                 ' but an object of class database was expected', 3)
+            
         results = dxpy.api.database_list_folder(
-            args.database,
+            entity_result['id'],
             input_params={"folder": args.folder, "recurse": args.recurse, "timeout": args.timeout})
         for r in results["results"]:
             date_str = render_timestamp(r["modified"]) if r["modified"] != 0 else ''
@@ -2963,7 +2981,7 @@ def run_body(args, executable, dest_proj, dest_path, preset_inputs=None, input_n
         "ignore_reuse_stages": args.ignore_reuse_stages or None,
         "debug": {"debugOn": args.debug_on} if args.debug_on else None,
         "delay_workspace_destruction": args.delay_workspace_destruction,
-        "priority": ("high" if args.watch or args.ssh or args.allow_ssh else args.priority),
+        "priority": args.priority,
         "instance_type": args.instance_type,
         "stage_instance_types": args.stage_instance_types,
         "stage_folders": args.stage_folders,
@@ -2974,6 +2992,16 @@ def run_body(args, executable, dest_proj, dest_path, preset_inputs=None, input_n
         "extra_args": args.extra_args
     }
 
+    if any([args.watch or args.ssh or args.allow_ssh]):
+        if run_kwargs["priority"] in ["low", "normal"]:
+            if not args.brief:
+                print(fill(BOLD("WARNING") + ": You have requested that jobs be run under " +
+                        BOLD(run_kwargs["priority"]) +
+                        " priority, which may cause them to be restarted at any point, interrupting interactive work."))
+                print()
+        else: # if run_kwargs["priority"] is None
+            run_kwargs["priority"] = "high"
+    
     if run_kwargs["priority"] in ["low", "normal"] and not args.brief:
         special_access = set()
         executable_desc = executable_describe or executable.describe()
@@ -4781,7 +4809,7 @@ parser_list_database_files = subparsers_list_database.add_parser(
     parents=[env_args],
     prog='dx list database files'
 )
-parser_list_database_files.add_argument('database', help='ID of the database.')
+parser_list_database_files.add_argument('database', help='Data object ID or path of the database.')
 parser_list_database_files.add_argument('--folder', default='/', help='Name of folder (directory) in which to start searching for database files. This will typically match the name of the table whose files are of interest. The default value is "/" which will start the search at the root folder of the database.')
 parser_list_database_files.add_argument("--recurse", default=False, help='Look for files recursively down the directory structure. Otherwise, by default, only look on one level.', action='store_true')
 parser_list_database_files.add_argument("--csv", default=False, help='Write output as comma delimited fields, suitable as CSV format.', action='store_true')
@@ -4991,7 +5019,7 @@ parser_run.add_argument('-h', '--help', help='show this help message and exit', 
 parser_run.add_argument('--clone', help=fill('Job or analysis ID or name from which to use as default options (will use the exact same executable ID, destination project and folder, job input, instance type requests, and a similar name unless explicitly overridden by command-line arguments. When using an analysis with --clone a workflow executable cannot be overriden and should not be provided.)', width_adjustment=-24))
 parser_run.add_argument('--alias', '--version', dest='alias',
                         help=fill('Alias (tag) or version of the app to run (default: "default" if an app)', width_adjustment=-24))
-parser_run.add_argument('--destination', '--folder', metavar='PATH', dest='folder', help=fill('The full project:folder path in which to output the results.  By default, the current working directory will be used.', width_adjustment=-24))
+parser_run.add_argument('--destination', '--folder', metavar='PATH', dest='folder', help=fill('The full project:folder path in which to output the results. By default, the current working directory will be used.', width_adjustment=-24))
 parser_run.add_argument('--batch-folders', dest='batch_folders',
                         help=fill('Output results to separate folders, one per batch, using batch ID as the name of the output folder. The batch output folder location will be relative to the path set in --destination', width_adjustment=-24),
                         action='store_true')
@@ -5019,19 +5047,20 @@ parser_run.add_argument('--delay-workspace-destruction',
                         action='store_true')
 parser_run.add_argument('--priority',
                         choices=['low', 'normal', 'high'],
-                        help='Request a scheduling priority for all resulting jobs. Will be overriden (set to high) ' +
-                             'when either --watch, --ssh, or --allow-ssh flags are used')
+                        help=fill('Request a scheduling priority for all resulting jobs. ' +
+                                  'Defaults to high when --watch, --ssh, or --allow-ssh flags are used.', 
+                                  width_adjustment=-24))
 parser_run.add_argument('-y', '--yes', dest='confirm', help='Do not ask for confirmation', action='store_false')
 parser_run.add_argument('--wait', help='Wait until the job is done before returning', action='store_true')
-parser_run.add_argument('--watch', help="Watch the job after launching it; sets --priority high", action='store_true')
+parser_run.add_argument('--watch', help="Watch the job after launching it. Defaults --priority to high.", action='store_true')
 parser_run.add_argument('--allow-ssh', action='append', nargs='?', metavar='ADDRESS',
-                        help=fill('Configure the job to allow SSH access; sets --priority high. If an argument is ' +
+                        help=fill('Configure the job to allow SSH access. Defaults --priority to high. If an argument is ' +
                                   'supplied, it is interpreted as an IP or hostname mask to allow connections from, ' +
                                   'e.g. "--allow-ssh 1.2.3.4 --allow-ssh berkeley.edu"',
                                   width_adjustment=-24))
 parser_run.add_argument('--ssh',
-                        help=fill("Configure the job to allow SSH access and connect to it after launching; " +
-                                  "sets --priority high",
+                        help=fill("Configure the job to allow SSH access and connect to it after launching. " +
+                                  "Defaults --priority to high.",
                                   width_adjustment=-24),
                         action='store_true')
 parser_run.add_argument('--ssh-proxy', metavar=('<address>:<port>'),