+++ src/python/dxpy/utils/describe.py
@@ -31,7 +31,7 @@ from collections import defaultdict
 
 import dxpy
 from .printing import (RED, GREEN, BLUE, YELLOW, WHITE, BOLD, UNDERLINE, ENDC, DELIMITER, get_delimiter, fill)
-from ..compat import basestring
+from ..compat import basestring, USING_PYTHON2
 
 def JOB_STATES(state):
     if state == 'failed':
@@ -756,10 +756,15 @@ def print_data_obj_desc(desc, verbose=False):
 def printable_ssh_host_key(ssh_host_key):
     try:
         keygen = subprocess.Popen(["ssh-keygen", "-lf", "/dev/stdin"], stdin=subprocess.PIPE, stdout=subprocess.PIPE)
-        (stdout, stderr) = keygen.communicate(ssh_host_key)
+        if USING_PYTHON2:
+            (stdout, stderr) = keygen.communicate(ssh_host_key)
+        else:
+            (stdout, stderr) = keygen.communicate(ssh_host_key.encode())
     except:
         return ssh_host_key.strip()
     else:
+        if not USING_PYTHON2:
+            stdout =  stdout.decode()
         return stdout.replace(" no comment", "").strip()